refactor: move localizeNumber to utils, restore localizeNumber signature, update imports
