+++ packages/generate-pdf-template-helpers/test/localize.spec.ts
@@ -1,4 +1,6 @@
-import { localizeDate, localizeNumber, localizeValue } from '../src';
+import { localizeNumber } from '@s1seven/schema-tools-utils';
+
+import { localizeDate, localizeValue } from '../src';
 
 describe('Localize', () => {
   const testNumber = 12000;
